feat: add Sort Colors (75) solution for Kaido

+++ dsa/leetcode/medium/Kaido/75.cpp
@@ -6,7 +6,7 @@ using namespace std;
 class Solution {
 public:
     void sortColors(vector<int>& nums) {
-        size_t l=0,m=0,h=nums.size()-1;
+        int l=0,m=0,h=nums.size()-1;
         while(m<=h)
             if(nums[m]==0)      swap(nums[m++],nums[l++]);
             else if(nums[m]==1) m++;
